Reject unsupported primitive storage before implicit byte lowering
Validate used local signatures, including address-written locals, rather than skipping every non-Int32 primitive. Reject unsupported static primitive fields consistently with the documented local-only Int32 compaction model. Keep allocation fixtures on actual supported short/ushort storage; retain identical emitted expectations and every golden. Twenty regression cases cover local, address, field and unused-slot boundaries. Debug and Release each pass 1291 tests.

+++ src/dotnes.tests/LocalsTests.cs
@@ -344,7 +344,7 @@ public class LocalsTests : ExecutionTests
             static class G
             {
                 public static byte byte_val;
-                public static int int_val;
+                public static short int_val;
             }
             """);
         var hex = Convert.ToHexString(bytes);
@@ -385,7 +385,7 @@ public class LocalsTests : ExecutionTests
             static class G
             {
                 public static byte byte_val;
-                public static int int_val;
+                public static short int_val;
                 public static ushort word_val;
             }
             """);
